medical_vqa/requirements.txt
torch>=2.0.0
transformers>=4.30.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=9.5.0
tqdm>=4.65.0
streamlit>=1.24.0
scikit-learn>=1.2.0
nltk>=3.8.0
evaluate>=0.4.0